boolean attributes and tests
